backend/requirements_hosting.txt
# Requirements.txt для виртуального хостинга (Python 3.6-3.8)
# Совместимые версии, которые работают на старых серверах

# Django LTS версия - стабильная и безопасная
django>=3.2,<3.3

# Django REST Framework - совместимая версия
djangorestframework>=3.12,<3.15

# База данных (PostgreSQL опционально, SQLite встроенно)
psycopg2-binary>=2.8,<2.10

# CORS заголовки
django-cors-headers>=3.10,<4.0

# Веб-сервер для production
gunicorn>=20.0,<22.0

# Статические файлы
whitenoise>=5.0,<7.0

# Обработка изображений
Pillow>=8.0,<11.0

# HTTP запросы
requests>=2.25,<3.0

# Богатый текстовый редактор (совместимая версия)
django-ckeditor>=6.0,<7.0

# Фильтрация
django-filter>=21.0,<24.0

# Переменные окружения
python-dotenv>=0.19,<2.0

# Дополнительные зависимости для совместимости
six>=1.15.0
pytz>=2021.1